+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Support Winnow that has outSR
+
 ## [1.3.2] - 2017-01-31
 ### Fixed
 * Recompiled